+++ src/RefDocGen/MemberData/FieldData.cs
@@ -1,5 +1,5 @@
-using RefDocGen.DocExtraction.Tools;
 using RefDocGen.MemberData.Abstract;
+using RefDocGen.Tools.Xml;
 using System.Reflection;
 using System.Xml.Linq;
